tools/discord-bot/requirements.txt
discord.py>=2.3.0
httpx>=0.24.0
python-dotenv>=1.0.0
